removed all unesuful questions and tested regeneration without new prompt
