+++ generators/languages/USAGE
@@ -1,0 +1,24 @@
+Description:
+    Select languages from a list of available languages. The i18n files will be copied to the /webapp/i18n folder.
+
+Example:
+    jhipster languages
+
+    This will create for each selected language:
+        /webapp/i18n/{lang}/global.json
+        /webapp/i18n/{lang}/activate.json
+        /webapp/i18n/{lang}/audits.json
+        /webapp/i18n/{lang}/configuration.json
+        /webapp/i18n/{lang}/error.json
+        /webapp/i18n/{lang}/gateway.json
+        /webapp/i18n/{lang}/health.json
+        /webapp/i18n/{lang}/home.json
+        /webapp/i18n/{lang}/login.json
+        /webapp/i18n/{lang}/logs.json
+        /webapp/i18n/{lang}/metrics.json
+        /webapp/i18n/{lang}/password.json
+        /webapp/i18n/{lang}/register.json
+        /webapp/i18n/{lang}/reset.json
+        /webapp/i18n/{lang}/sessions.json
+        /webapp/i18n/{lang}/settings.json
+        /webapp/i18n/{lang}/user-management.json
